Refactor service layer to have abstract classes
